+++ tests/testthat/test-gam.R
@@ -1,9 +1,11 @@
 .runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
 
 if (.runThisTest) {
-  if (require("testthat") && require("insight") && require("mgcv")) {
+  if (requiet("testthat") && requiet("insight") && requiet("mgcv")) {
     set.seed(123)
-    dat <- mgcv::gamSim(1, n = 400, dist = "normal", scale = 2)
+    void <- capture.output(
+      dat <- mgcv::gamSim(1, n = 400, dist = "normal", scale = 2)
+    )
     m1 <- mgcv::gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
 
     m2 <- download_model("gam_zi_1")
@@ -191,5 +193,27 @@ if (.runThisTest) {
     test_that("find_statistic", {
       expect_identical(find_statistic(m1), "t-statistic")
     })
+
+    test_that("get_predicted", {
+      tmp <- mgcv::gam(y ~ s(x0) + s(x1), data = head(dat, 30))
+      pred <- get_predicted(tmp)
+      expect_s3_class(pred, "get_predicted")
+      expect_snapshot(print(pred))
+
+      x <- get_predicted(tmp, predict = NULL, type = "link")
+      y <- get_predicted(tmp, predict = "link")
+      z <- predict(tmp, type = "link", se.fit = TRUE)
+      expect_equal(x, y)
+      expect_equal(x, z$fit, ignore_attr = TRUE)
+      expect_equal(as.data.frame(x)$SE, z$se.fit, ignore_attr = TRUE)
+
+      x <- get_predicted(tmp, predict = NULL, type = "response")
+      y <- get_predicted(tmp, predict = "expectation")
+      z <- predict(tmp, type = "response", se.fit = TRUE)
+      expect_equal(x, y, ignore_attr = TRUE)
+      expect_equal(x, z$fit, ignore_attr = TRUE)
+      expect_equal(as.data.frame(x)$SE, z$se.fit, ignore_attr = TRUE)
+    })
+
   }
 }